backend/requirements.txt
flask
python-dotenv
openai
faiss-cpu
numpy
flask-cors
gunicorn
Flask-SQLAlchemy
Flask-Migrate